🗑️ cleanup: Organize folder and file structure
