+++ app/(conversation)/_layout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from 'expo-router';
+
+export default function Layout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+      }}
+    />
+  );
+}
