+++ app/_layout.tsx
@@ -75,6 +75,7 @@ const InitialLayout = () => {
     <Stack>
       <Stack.Screen name="(auth)" options={{ headerShown: false }} />
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      <Stack.Screen name="(conversation)" options={{ headerShown: false }} />
       <Stack.Screen name="+not-found" />
     </Stack>
   );
